initial UserForm; refactor UserEdit
