+++ src/pages/Users/components/UserEdit/UserEdit.scss
@@ -1,0 +1,9 @@
+.user-edit {
+  .row-message {
+    margin: 1rem 0;
+
+    &.row-card .wrapper {
+      width: 100% !important;
+    }
+  }
+}
