--- src/pages/Users/components/UserEdit/UserEditForm.scss
@@ -1,32 +1,0 @@
-.form-user-edit {
-  max-width: 45rem;
-
-  margin: 1rem 0;
-
-  ion-input {
-    margin-bottom: 0.5rem;
-  }
-
-  .buttons {
-    display: flex;
-
-    ion-button {
-      width: 100%;
-    }
-
-    .spinner-button-save {
-      ion-spinner {
-        height: 16px;
-        width: 16px;
-      }
-    }
-  }
-
-  .row-message {
-    margin: 1rem 0;
-
-    &.row-card .wrapper {
-      width: 100% !important;
-    }
-  }
-}
