+++ src/pages/Users/components/UserForm/UserForm.scss
@@ -1,0 +1,13 @@
+.form-user {
+  ion-input {
+    margin-bottom: 0.5rem;
+  }
+
+  .buttons {
+    display: flex;
+
+    ion-button {
+      width: 100%;
+    }
+  }
+}
